updated script

+++ scripts/snowbridge_update_subtree.sh
@@ -42,7 +42,7 @@ function fetch() {
 function clean() {
     echo ""
     echo "Patching/removing unneeded stuff from subtree in target directory: '$SNOWBRIDGE_TARGET_DIR'"
-    #remove_parachain_dir
+    remove_parachain_dir
     $SNOWBRIDGE_TARGET_DIR/scripts/verify-pallets-build.sh --ignore-git-state --no-revert
 }
 
@@ -53,26 +53,34 @@ function create_patch() {
         exit 1;
     }
     echo "Creating diff patch file to apply to snowbridg. No Cargo.toml files will be included in the patch."
-    #add_parachain_dir
+    add_parachain_dir
     git diff snowbridge/$SNOWBRIDGE_BRANCH $POLKADOT_SDK_BRANCH:bridges/snowbridge --diff-filter=ACM -- . ':(exclude)*/Cargo.toml' > snowbridge.patch
+    remove_parachain_dir
 }
 
 function remove_parachain_dir() {
     SOURCE_DIR="bridges/snowbridge/parachain"
     TARGET_DIR="bridges/snowbridge"
+    if [[ -e $SOURCE_DIR ]]; then
+        echo "Removing parachain dir"
+        rm -r $TARGET_DIR/scripts # because there's scripts in both dirs
+        mv $SOURCE_DIR/* $TARGET_DIR/
 
-    mv $SOURCE_DIR/* $TARGET_DIR/
-
-    rm -r $SOURCE_DIR
+        rm -r $SOURCE_DIR
+    else
+        echo "Parachain dir already cleared"
+    fi
 }
 
 function add_parachain_dir() {
     SOURCE_DIR="bridges/snowbridge"
     TARGET_DIR="bridges/snowbridge/parachain"
-
-    mkdir -p $TARGET_DIR
-
-    mv $SOURCE_DIR/* $TARGET_DIR/
+    if [[ -e $SOURCE_DIR ]]; then
+        echo "Parachain dir already added"
+    else
+        mkdir -p $TARGET_DIR
+        mv $SOURCE_DIR/* $TARGET_DIR/
+    fi
 }
 
 case "$1" in
